+++ Client/GameCoding/Utils.cpp
@@ -1,7 +1,6 @@
 #include "pch.h"
 #include "Utils.h"
 
-
 // static_cast로 번거롭게 정수로 한 이유는?
 // 세밀한 움직임을 구현하기 위해서 float로 만들어놨기 때문
 
@@ -12,7 +11,7 @@ void Utils::DrawText(HDC hdc, Pos pos, const wstring& str)
 
 void Utils::DrawRect(HDC hdc, Pos pos, int32 w, int32 h)
 {
-	::Rectangle(hdc, static_cast<int32>(pos.x - w / 2), static_cast<int32>(pos.y - h / 2), static_cast<int32>(pos.x + w / 2), static_cast<int32>(pos.y + h / 2));
+	::Rectangle(hdc, static_cast<int32>(pos.x - (float)(w / 2)), static_cast<int32>(pos.y - (float)h / 2), static_cast<int32>(pos.x + (float)w / 2), static_cast<int32>(pos.y + (float)h / 2));
 }
 
 void Utils::DrawCircle(HDC hdc, Pos pos, int32 radius)
